Refactor createTable logic into CassandraDataStore
